Silicon/NXP/I2cLib: Add Pcd for early I2cInit
The I2cDivisor is determined like this
I2cDivisor = (I2cClock + I2cSpeed - 1) / I2cSpeed
I2cClock is the input clock to I2c controller, I2cSpeed is the I2c bus
speed. In early init phase when we don't know the I2cClock, we need to
setup the I2c controller to be able to read SysClock (and hence I2cClock)
from the clock generator or FPGA connected to I2c bus.
Therefore, this Pcd selects the Divisor to use in early init phase.

+++ Silicon/NXP/Library/I2cLib/I2cLibInternal.h
@@ -80,7 +80,7 @@
 
 typedef struct _I2C_REGS {
   UINT8 Ibad; // I2c Bus Address Register
-  UINT8 Ibfd; // I2c Bus Frequency Dividor Register
+  UINT8 Ibfd; // I2c Bus Frequency Divider Register
   UINT8 Ibcr; // I2c Bus Control Register
   UINT8 Ibsr; // I2c Bus Status Register
   UINT8 Ibdr; // I2C Bus Data I/O Register
@@ -93,7 +93,7 @@ typedef struct _I2C_REGS {
  */
 typedef struct _I2C_CLOCK_DIVISOR_PAIR {
   UINT16  Divisor;
-  UINT16  Ibfd; // I2c Bus Frequency Dividor Register value
+  UINT8   Ibfd; // I2c Bus Frequency Divider Register value
 } I2C_CLOCK_DIVISOR_PAIR;
 
 typedef struct {
